Prototype/Backend/requirements.txt
flask
werkzeug
flask_bootstrap
flask_sqlalchemy
flask_cors
flask_login
flask_mail
psycopg2
bcrypt
pyjwt
itsdangerous